Move code from GameBuilder -> GameWindow

+++ src/GameArea.h
@@ -22,7 +22,7 @@ private:
 
     /* declare counter used to know which of the flag imgs
      * which compose the flag's animation should be showed */
-    int flagCounter;
+    unsigned long flagCounter;
 
     /* DRAWING METHODS */
     void drawFlagAnimation(const Cairo::RefPtr<Cairo::Context> &ptr,
